+++ src/AssociatedAsset.php
@@ -3,44 +3,13 @@
 namespace Thinktomorrow\AssetLibrary;
 
 use Illuminate\Database\Eloquent\Relations\MorphPivot;
-use Illuminate\Support\Arr;
 
 class AssociatedAsset extends MorphPivot
 {
+    use ProvidesData;
+
     public $guarded = [];
     protected $casts = [
         'data' => 'array',
     ];
-
-    public function hasData(string $key): bool
-    {
-        return Arr::has($this->data, $key);
-    }
-
-    public function getData(string $key, $default = null)
-    {
-        return Arr::get($this->data, $key, $default);
-    }
-
-    public function setData(string $name, $value): self
-    {
-        $data = $this->data;
-
-        Arr::set($data, $name, $value);
-
-        $this->data = $data;
-
-        return $this;
-    }
-
-    public function forgetData(string $name): self
-    {
-        $data = $this->data;
-
-        Arr::forget($data, $name);
-
-        $this->data = $data;
-
-        return $this;
-    }
 }