+++ CHANGELOG.md
@@ -7,7 +7,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 ## unreleased
 
 ## Upgrade to 0.9
--   Added: it is now possible to manage data per attached asset. 
+-   Added: it is now possible to manage data per asset or per associated pivot record. 
 -   Added: You can now define additional formats in which the asset should be converted.
 -   Changed: (Breaking) Rename `Thinktomorrow\AssetLibrary\AssetTrait` to `Thinktomorrow\AssetLibrary\InteractsWithAssets`.
 -   Changed: Following methods are deprecated: Asset::filename(), Asset::url() and Asset::hasFile(). Use Asset::getFileName(), Asset::getUrl() and Asset::exists() instead.